Fix missing PXE directories for Conductor
When Ironic Conductor class is called, it expects the
PXE directories exist. That is only the case for the step 4.
While there is also a case when the conductor class invoked
for the step 3 & db sync case. For that case also inlcude
the missing ironic::pxe class to ensure the PXE directories
created.

Closes-Bug: #1845222

Change-Id: I394f56ba9b213c75378bdf21999d23509632523c
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>