Allow to use the upstream rabbitmq-server-ha OCF resource agent
We introduce a new hiera key in order to be able to use the upstream
rabbitmq-server-ha OCF resource [1].
For it to work inside bundles we need to have a rabbitmq-server
package inside the bundle which includes at least
https://github.com/rabbitmq/rabbitmq-server/pull/2853
and also we need to be using at least pacemaker-2.0.4-4.el8.

The rationale for this work is that the current rabbitmq-cluster
resource agent maintained under the ClusterLabs umbrella is
a cloned resource, which is limited in the number of actions
it can do in a number of situations (partition, failover, etc).

The upstream resource agent is a master/slave resource and
allows for more expressive semantics in general and is preferrable [2].

[1] https://github.com/rabbitmq/rabbitmq-server/blob/master/scripts/rabbitmq-server-ha.ocf
[2] https://github.com/lemenkov/pmk-rmq.md/blob/master/pmk-rmq.md

Co-authored-by: Bogdan Dobrelya <bdobrelia@redhat.com>
Change-Id: Ia273d0dbc668bbae4c6e9cb535bd68783faf0148