pacemaker: add support for Hash vs List in container environment
We are transitioning from an array to an hash for the container
environment of each container:
I894f339cdf03bc2a93c588f826f738b0b851a3ad

Mainly to make it consummable by Ansible later; where the
podman_container module needs a dict instead of a list.

This patch just changes the default, and also adds support for an Hash
instead of a List, but still supporting the List.

Change-Id: I4e53a4a3464940660473bcbe74e30507a69a4019