Improve stonith leves idempotency.
This commit improves the way stonith levels are set up and their
resiliency against redeployments by introducing a stonith_levels
custom fact that collects the current stonith levels defined for
the specific server, so we can compare against the desired number
of levels defined in hiera.

If these do not match (for example if there are additional levels
that are no longer necessary), the clean up step also introduced
by this commit takes care of deleting the ones no longer necessary.

Change-Id: Ifae73ac2bf4481d0a11e89c0ea0916e85dd2db1d