Avoid failing on rsyslog
This patch makes deployment pass when value of hiera variable
tripleo_logging_sources_<enabled_service> is undef. This case should
technically not happen, but currently enabled_services contains
services which are not being deployed, so we need to make sure that
Puppet does not fail in this case.

Change-Id: Ib34304dada25a0be1a39f508119adbe9cca4d66f
(cherry picked from commit 1207666e173543b63fbd76728553b31274399576)