Add short node name override support to manila
In line with the cinder & cinder-backup services, we add
short node name override support to manila when implemented
with pacemaker.

This is required in order to complete the 'scale-up' upgrade
steps for manila similar to the implementation in
I3cfd4272449b207874723682903505d2bbdb435b

Included is also an update to the tests and their fixtures to
ensure that the unit tests for the manila data are consistent
with the other similar services (cinder volume and cinder
backup).

Related-Bug: #1838971
Change-Id: I2c468e608a8f70d031b9ba08521bb75f7094c048
(cherry picked from commit e1348fecc80a5c746fd612d7c8a3cbbb58a974d4)