Remove unused manila_share_nodes_count variable
The manila_share_nodes_count variable is defined, but not used
anywhere. In some services this is used for the replica count,
but manila (similar to cinder) has a hard-coded replica count
of 1 and therefore this variable is not needed.

Change-Id: I6398871a0a8c235e626748206dda617dadb06895
(cherry picked from commit 9ca6d5a1943e6be6f81b1a1dda94fdc071f043b6)