Change the way HAProxy service is reloaded
NOTE: In queens, this changes the way haproxy is reloaded to the
same way it is reloaded in files/certmonger-haproxy-refresh.sh,
since there is no systemd unit file for the service.

The master patch only changes the name of the haproxy service
so that it reflects the new name of the service.

Resolves: rhbz#1698058

Change-Id: I66ee58f3b4fd2f10ceba6306497ac796daaf98e8
(cherry picked from commit 48125267ddb8e2a14073a414725f55b314e31134)