Add unit test job on CentOS8
This patch introduces a unit test job which runs on CentOS8, so that we
can migrate from CentOS7 and CentOS8.
The new unit test job on CentOS8 is added as non-voting job initially,
but will be make as voting job with removing CentOS7 one after we
confirm the gate status of master and stable branches after this change
is merged.

Change-Id: Iee33fe1953af27b5f4b68b093464a831cb4ddcc6