Update metadata to support el8 instead of el7 again
This patch again updates metadata to support el8 instead of el7, which
was initially merged[1] but reverted because of failures in unit tests
caused by old puppet modules installed as dependencies.

This patch removes version cap for puppet-etcd and puppet-redis, so
that we can use their lastet verions compatible with el8.

[1] ff22e7c2687bc8f67d6769c32b235257ac0f8e5f

Related-Bug: #1872562
Change-Id: I3924d35fae48accb57186c1c5c0a3192bacccda6