Hardcode openidc_cache_type parameter for keystone federation with OpenIdc
By default OpenIdc uses shared memory caching mode, which will
not work for multiple controller nodes. puppet-tripleo already
configured to calculate memcached servers for OpenIdc, but for
some reason doesn't set "openidc_cache_type" to "memcache", so
shm is used.

There are a number of options available for "openidc_cache_type",
but memcache is the only one that will currently work for multiple
controllers:

- shm and file are stored locally on every node;
- redis requires mod_auth_openidc to be compiled with redis
  support, but it is not generally the case.

To avoid providing illusion of freedom of choice, it would
be right to hardcode this in puppet-tripleo.

Closes-Bug: #1873239
rhbz: #1824506

Change-Id: I7cbc462b2ff99b7b0d3ff58fda1b52ccf85fc86d