Fix RABBITMQ_SERVER_ADDITIONAL_ERL_ARGS usage with a proper override mechanism
In THT we allow RabbitAdditionalErlArgs to set some additional
parameters which should be passed to RABBITMQ_SERVER_ADDITIONAL_ERL_ARGS
when starting rabbit. The problem is that when we use internal tls
that parameter gets ignored and so we lose our default '+sbwt none'.

Let's do this via a proper parameter while also considering the fact
that historically the default value of RabbitAdditionalErlArgs had
apices around it.

Change-Id: I3bf244a70538209773804eb85fae6be035c587f4
Related-Bug: #1884922