Fix up cluster_setup_extras usage
Starting with pcs 0.10 (CentOS/RHEL 8) the arguments to pcs cluster
setup changed completely in an incompatible way and in fact are ignored
by puppet-pacemaker. As we would like to reenable them, we first need to
make sure that whatever we pass is not understood by pcs.

We can also drop ipv6 and encryption parameters as corosync/pcs do not
need them any longer.

Change-Id: Iaecced845b80e958c7ad0ef96896f1202259cca2
Related-Bug: #1886789