Use correct memcached IPs for Swift object-expirer
The memcached IP is not properly set in case of separate storage nodes.
Until now, these were only set if the Swift proxy node was also
configured on the same node; however, this is not the case with separate
storage nodes. The service would still run, but was not able to use any
cache and logging errors therefore.

Closes-Bug: 1889663
Change-Id: I70ac9fcd3f74975c96c715ecbc818e6606ff4517