Fix tls - zaqar wss default port
In https://review.opendev.org/746671 we updated the
endpoint map. It seems this broke the undercloud as
haproxy still listen on port 9000.

Update the default zaqar wss port in puppet as well.

Also set the request header Host to use the service
port so that handshake does not fail on port missmatch.

Change-Id: I0c9dd4e9457c7cfa8cd54b5eaa167bf4e8e6bcc1
Closes-Bug: #1893621
Related-Bug: #1832168
Related-RHBZ: #1868910