[Train and before] Apply default_listen_options to ironic-inspector
93ae3c95955e0c66550d3cf50f3d1b8715a935e5 introduced additional
settings to listen_options for ironic-inspector endpoint, but it didn't
merge default_listen_options but removed the default settings
completely.

This patch make sure that default settings are merged, so that some
important settings like X-Forwarded-Proto header is applied correctly.

Note that the above patch is reverted during Ussuri cycle, thus we need
this change only in Train and before.

Change-Id: I5877c07be84b8f10b4ac125ef9e22c40d2fcac0d