Resolve conflict during module installation for unit tests
Since the change[1] in p-o-i is merged, selinux_core is automatically
installed during unit test and inistallation of modules fails if we
have the same module included in Puppet_extras file.

This patch removes selinux_core from Puppet_extras because for
the current CI jobs of puppet-tripleo it's enough if the package is
somehow installed during unit tests.

[1] ef3c5dc9b7294b38b95c4e900bd142987472f976

Closes-Bug: #1894984
Change-Id: I15e541a641969af7ac117faf13d67827662e7ca6