Add changes for collectd libpod stats plugin
Because the libpod stats plugin is written in golang rather than C,
it is not included in the collectd rpm and therefor cannot be
configured by puppet-collectd and are made here instead. A simple
variable for enabling will suffice for now since the plugin does
not currently have any configuration options.

Change-Id: I575b6504e24f9d690e1b88b2a6bec296438ce747
Signed-off-by: pleimer <pfbleimer@gmail.com>