Make nova_metadata use source LB strategy
Generating the metadata for a VM that has been freshly booted
is a fairly expensive operation. Let's use a source balance option
in the nova_metadata haproxy stanza so that the requests from
the same compute node end up on the same controller, hence increasing
the locality of this operation.

Generated haproxy stanza looks like:
listen nova_metadata
  bind 172.17.1.150:8775 transparent ssl crt /etc/pki/tls/certs/haproxy/overcloud-haproxy-internal_api.pem
  mode http
  balance source
  hash-type consistent
  http-request set-header X-Forwarded-Proto https if { ssl_fc }
  http-request set-header X-Forwarded-Proto http if !{ ssl_fc }
  http-request set-header X-Forwarded-Port %[dst_port]
  option httpchk
  option httplog
  server controller-0.internalapi.foo.local 172.17.1.90:8775 ca-file /etc/ipa/ca.crt check fall 5 inter 2000 rise 2 ssl verify required verifyhost controller-0.internalapi.foo.local
  server controller-1.internalapi.foo.local 172.17.1.95:8775 ca-file /etc/ipa/ca.crt check fall 5 inter 2000 rise 2 ssl verify required verifyhost controller-1.internalapi.foo.local
  server controller-2.internalapi.foo.local 172.17.1.87:8775 ca-file /etc/ipa/ca.crt check fall 5 inter 2000 rise 2 ssl verify required verifyhost controller-2.internalapi.foo.local

We also add the consistent hash-type as it should improve the cache
behaviour when a controller disappears or gets added.

Change-Id: I4847ae10b03cf7b0571976097262bc8069c4e070