Allow to specify a nic for the VIPs
This review allows the operator to specify a nic to bind to a VIP to.
It does so in two different ways:
1) Via a global 'tripleo::pacemaker::force_nic' hiera key.
For example setting 'tripleo::pacemaker::force_nic: lo'
will make sure all VIPs are added to the 'lo' network interface.

2) The global setting can be overriden via a custom hiera
hash called 'force_vip_nic_overrides' so that we can
override the binding nic for specific VIPs. For example:

  force_vip_nic_overrides:
    redis_vip: vip-test-nic
    ovn_dbs_vip: vip-test-nic

Will make sure that the redis_vip and the ovn_dbs_vip will
be bound to the 'vip-test-nic'.

Tested with:
  tripleo::pacemaker::force_nic: lo
  force_vip_nic_overrides:
    redis_vip: vip-test-nic
    ovn_dbs_vip: vip-test-nic

And correctly got all VIPs running on lo:
1: lo    inet 192.168.24.9/32 scope global lo\       valid_lft forever preferred_lft forever
1: lo    inet 172.23.1.9/32 scope global lo\       valid_lft forever preferred_lft forever

except for redis and ovn_dbs which are on the vip-test-nic:
5: vip-test-nic    inet 172.25.1.7/32 scope global vip-test-nic\       valid_lft forever preferred_lft forever

Note: The interface is intentionally kept in hiera as this change
will only be needed whenever BGP runs on all nodes and advertises the
VIPs across the ASN (i.e. we do not want to make it very exposed just
yet)

Change-Id: Iaa2676655d97a66237a3bd6066985f87d3565a4d