Fix nic selection when no nic is specified
This is a way to reinstate change
https://review.opendev.org/c/openstack/puppet-tripleo/+/761833 which was
supposed to remove a workaround for VIP selection on IPv6. Turns out
the workaround is still needed because the resource-agent for IPaddr2
is still problematic with IPv6:
https://bugzilla.redhat.com/show_bug.cgi?id=1902851

With this change we still force a nic when using ipv6 except in the
case a user specifies the nic via the tripleo::pacemaker::force_nic
hiera key.

Tested this and I can correctly bind the VIPs as expected:
    tripleo::pacemaker::force_nic: lo
    force_vip_nic_overrides:
      redis_vip: vlan20
      ovn_dbs_vip: vlan20

[root@controller-2 ~]# grep -ir ovn_dbs_v /etc/puppet/hieradata/
/etc/puppet/hieradata/all_nodes.json:    "ovn_dbs_vip": "172.17.1.107",
/etc/puppet/hieradata/extraconfig.json:        "ovn_dbs_vip": "vlan20",
[root@controller-2 ~]# ip -o a |grep 17.1.107/32
10: vlan20    inet 172.17.1.107/32 brd 172.17.1.255 scope global vlan20\
    valid_lft forever preferred_lft forever

Change-Id: I14bfaf0060093811cddf0cdd9a130b91a3cd0477