Allow to specify a nic for the VIPs + Fix nic selection when no nic is specified
This review is a merge of:
Revert of Revert "Cleanup old workaround for ipv6 VIPs": Ia41046f148e0593ea773e8409494ce5dcca4b7a2
Fix nic selection when no nic is specified : I14bfaf0060093811cddf0cdd9a130b91a3cd0477
Allow to specify a nic for the VIPs : Iaa2676655d97a66237a3bd6066985f87d3565a4d↲

We do this merge only for ussuri and train because in master and
victoria we landed "Allow to specify a nic for the VIPs" before doing
the revert of the "Cleanup old workaround for ipv6 VIPs".

So this change restores the exact state as we have in master and
victoria with a single review. We do not split the reviews because
if the Ia41046f148e0593ea773e8409494ce5dcca4b7a2 merges without
the other two we will have ipv6 broken.

This review allows the operator to specify a nic to bind to a VIP to.
It does so in two different ways:
1) Via a global 'tripleo::pacemaker::force_nic' hiera key.
For example setting 'tripleo::pacemaker::force_nic: lo'
will make sure all VIPs are added to the 'lo' network interface.

2) The global setting can be overriden via a custom hiera
hash called 'force_vip_nic_overrides' so that we can
override the binding nic for specific VIPs. For example:

  force_vip_nic_overrides:
    redis_vip: vip-test-nic
    ovn_dbs_vip: vip-test-nic

Will make sure that the redis_vip and the ovn_dbs_vip will
be bound to the 'vip-test-nic'.

Tested with:
  tripleo::pacemaker::force_nic: lo
  force_vip_nic_overrides:
    redis_vip: vip-test-nic
    ovn_dbs_vip: vip-test-nic

And correctly got all VIPs running on lo:
1: lo    inet 192.168.24.9/32 scope global lo\       valid_lft forever preferred_lft forever
1: lo    inet 172.23.1.9/32 scope global lo\       valid_lft forever preferred_lft forever

except for redis and ovn_dbs which are on the vip-test-nic:
5: vip-test-nic    inet 172.25.1.7/32 scope global vip-test-nic\       valid_lft forever preferred_lft forever

Note: The interface is intentionally kept in hiera as this change
will only be needed whenever BGP runs on all nodes and advertises the
VIPs across the ASN (i.e. we do not want to make it very exposed just
yet)

NB: Small conflict during chrry-pick due to context differences

Change-Id: I0630ae124dac304614431e67975767206555e9aa
(cherry picked from commit d632847311341f507f77cc2e9d9c5c0d0691c1b4)