Glance: Include the required classes for image cache
This change is a prep work to migrate class composition for image cache
feature from tripleo-heat-templates to puppet-tripleo, so that we can
gather all logics to compose required puppet classes in puppet-tripleo.

NOTE: We are not backporting db purge changes with this change because
the support of db purge cron job is not available in stable/train.

Depends-on: https://review.opendev.org/c/openstack/tripleo-quickstart/+/859516
Change-Id: I843d72542154a2e278ba257f6b61ed573c7c3860
(cherry picked from commit 950275ed0575e289da53642dba8bc03e1485416e)