Fix manila-share with ceph-nfs backend
Currently when you deploy manila-share with a ceph-nfs backend
we get the following error:
Error: pcs -f
/var/lib/pacemaker/cib/puppet-cib-backup20210107-48280-1y4rwmx
constraint colocation add openstack-manila-share with ceph-nfs INFINITY
failed: Error: Resource 'ceph-nfs' does not exist. Too many tries

The reason for it is that we moved the pcs resource creation to the host
and outside containers
https://review.opendev.org/q/topic:%22pcs_host3%22+(status:open%20OR%20status:merged)

But when the ceph-nfs pcmk resource gets created at step 5, the
manila-share puppet code has already been invoked and since it requires
the ceph-nfs resource to be available it fails like above.

Let's move the ceph-nfs creation to step4 to solve this.

Closes-Bug: #1910732

Change-Id: Ifb81d155cbb4598a311c8e6fe90be951a8e7e656