[train-only] certmonger: track change in service's private keys
Certmonger_certificate resources are marked as changed
by puppet one some of its properties changed, like e.g.
the filename of the service's private key.

However puppet-certmonger has no property to track that
the content of the private key has changed, so puppet
cannot trigger a certificate renewal when a user
explicitely regenerated this service's private key.

Add a file resource for tracking any content change in
the service's private key file, and make sure that puppet
notify puppet-certmonger to trigger a certificate renewal.

Note: some services that already use a file resource for
their private key are updated, because they configure a
wrong/backward dependency that causes a dependency loop
when we add a notify dependency to the file resource.

Related-Bug: #1906505

Change-Id: I48658413f69a68dcad0a2f24ea66fe027e987f26