Swift: Include puppet classes for gatekeeper and listing_formats
The gatekeeper middleware and the listing_formats middleware are always
enabled in proxy-server pipeline, even if these middlewares are not
explicitly enabled in the pipeline parameter.
This change ensure that the corresponding puppet classes are included
so that operators can set parameters of the these two middlewares.

This change also updates order of class inclusions so that the order
becomes consistent with the pipeline parameter.

Depends-on: https://review.opendev.org/776826
Change-Id: I994e549c5ccc9b06af2682826ac040edf51a3efa