Fix novavnc haproxy network
Since "Use single NovaLibvirtNetwork to configure instance console
components" Icef2481b65b41b524ad44eeecfbee4451006e1d2, we are getting
the following error:

Error while evaluating a Function Call, The service_network for this
service is undefined. Can't configure TLS for the internal network.

Let's first lookup the old network and then the newer one to keep
compatibility.

Tested by triggering fs039 on
https://review.opendev.org/c/openstack/tripleo-heat-templates/+/780934
with a dep-on on this review.

Change-Id: Ic9ae8a1a0f69d85b2acd2598644c1cae8c1db78d
Closes-Bug: #1919506