cinder: remove ineffective implementation about key_mgr_backend
Change c376d64e5f0b38165fa50feee04c2de0e1d851bc deprecated
the key_mgr_backend parameter of the cinder::api class, but the logic
to pick up the deprecatd parameter, which was implemented by this
change, was invalid because the cinder class is included before
the cinder::api class is included.
This change removes the invalid implementation and makes that parameter
just ineffective, because we use the cinder::* parameter from t-h-t
and it is unlikely that users would use that puppet-tripleo parameter
directly.

Change-Id: Ie24f82fcede4d2862f9e6a49734959499d9fd845