Add haproxy prometheus/alertmanager endpoint for Ceph Dashboard
This change adds the prometheus and alertmanager endpoints which
are on the same grafana network and allow us to just use a single
vip to access the prometheus and alertmanager instances from the
ceph dashboard when enabled.
We already did it for grafana [1] and we would like to do the same
for prometheus and alertmanager.

[1] https://github.com/ceph/ceph-ansible/blob/master/roles/ceph-dashboard/tasks/configure_dashboard.yml#L192

Change-Id: I0dbcde302a3abdfccc3dd996b1c238e25064dd12