stable-only: Pin puppetlabs-haproxy
The puppetlabs-haproxy repo no longer has the master branch and
the master branch has been replaced by the main branch.
While we switched the branch in use in master[1], for stable branches
it is better to pin it to a specific version to avoid issue caused by
any future change in puppetlabs-haproxy.

The version is determined based on the version currently provided in
RDO.

[1] https://review.opendev.org/c/openstack/puppet-tripleo/+/786018/4/Puppetfile_extras
Change-Id: I5d1d891474e4d7b47f497c5bd0144ce24e291260