Revert "Allow ovn_db_host to be of type array"
We shouldn't override the hieradata but rather be explicit about which
mode we're operating in. e.g. pacemaker or not. This will help in
troubleshooting and not rely on hieradata overrides to work.

This reverts commit 0b51d8258c4a0618ec6f2abd8f86bfe6943cd51f.

Related-Bug: #1931133
Change-Id: Id8df4fe5ce26dbcdeb7ba073ecb2199cf392b03b
(cherry picked from commit fad7e3fb90f2f03e011fd245e5b60d510726d9e3)